Refactor: Update AccountSettings, ChangePassword, and EditProfile components
